+++ src/main/scala/com/taxis/etl/extract/Downloader.scala
@@ -3,13 +3,11 @@ package com.taxis.etl.extract
 import java.nio.file.{Files, Paths}
 
 object Downloader {
-  def downloadFiles(download: IDownload): Unit = {
-    def folder: String = ReadFromEnv.readDownloadFolder()
+  def downloadFiles(download: IDownload, folder: String): Unit = {
     def path = Paths.get(folder)
-    if (!(Files.exists(path) && Files.isDirectory(path)))
+    if (!(Files.exists(path) && Files.isDirectory(path))) {
       Files.createDirectory(path)
-    ReadFromEnv.readTaxisETLFileName().foreach((filename: String) => {
-      download.perform(filename)
-    })
+      download.perform(ReadFromEnv.readTaxisETLFileName());
+    }
   }
 }
