+++ src/main/scala/com/taxis/etl/extract/IDownload.scala
@@ -1,7 +1,6 @@
 package com.taxis.etl.extract
 
-import java.nio.file.{Files, Paths}
 import scala.language.postfixOps
 trait IDownload {
-    def perform(filename: String): String
+    def perform(filenames: Array[String]): Unit
 }
