+++ src/main/scala/com/taxis/etl/extract/URLDownload.scala
@@ -6,8 +6,10 @@ import java.io.File
 
 import scala.language.postfixOps
 class URLDownload(baseUrl: String) extends IDownload {
-  def perform(filename: String): String = {
-    System.out.println("Downloading " + filename + " from " + baseUrl)
-    new URL(baseUrl + "/" + filename) #> new File(filename) !!
+  def perform(filenames: Array[String]): Unit = {
+    filenames.foreach((filename: String) => {
+      System.out.println("Downloading " + filename + " from " + baseUrl)
+      new URL(baseUrl + "/" + filename) #> new File(filename) !!
+    })
   }
 }
